requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
keyboard>=0.13.5
mouse>=0.7.1
pytesseract>=0.3.10
PyGetWindow>=0.0.9
pynput>=1.7.6
